Refactored localeCompare in SuggestionUtils
