+++ src/libs/SuggestionUtils.ts
@@ -1,6 +1,6 @@
+import type {LocaleContextProps} from '@components/LocaleContextProvider';
 import CONST from '@src/CONST';
 import type {PersonalDetails} from '@src/types/onyx';
-import localeCompare from './LocaleCompare';
 import {getDisplayNameForParticipant} from './ReportUtils';
 
 /**
@@ -32,19 +32,21 @@ function getDisplayName(details: PersonalDetails) {
 }
 
 /**
- * Comparison function to sort users. It compares weights, display names, and accountIDs in that order
+ * Function to sort users. It compares weights, display names, and accountIDs in that order
  */
-function compareUserInList(first: PersonalDetails & {weight: number}, second: PersonalDetails & {weight: number}) {
-    if (first.weight !== second.weight) {
-        return first.weight - second.weight;
-    }
+function getSortedPersonalDetails(personalDetails: Array<PersonalDetails & {weight: number}>, localeCompare: LocaleContextProps['localeCompare']) {
+    return personalDetails.sort((first, second) => {
+        if (first.weight !== second.weight) {
+            return first.weight - second.weight;
+        }
 
-    const displayNameLoginOrder = localeCompare(getDisplayName(first), getDisplayName(second));
-    if (displayNameLoginOrder !== 0) {
-        return displayNameLoginOrder;
-    }
+        const displayNameLoginOrder = localeCompare(getDisplayName(first), getDisplayName(second));
+        if (displayNameLoginOrder !== 0) {
+            return displayNameLoginOrder;
+        }
 
-    return first.accountID - second.accountID;
+        return first.accountID - second.accountID;
+    });
 }
 
-export {trimLeadingSpace, hasEnoughSpaceForLargeSuggestionMenu, compareUserInList};
+export {trimLeadingSpace, hasEnoughSpaceForLargeSuggestionMenu, getSortedPersonalDetails};
